admin: Split nav into primary + secondary links
