+++ admin/src/App.vue
@@ -1,6 +1,10 @@
 <template>
     <v-app>
-        <Navigation v-if="authService.isLoggedIn()" :items="navLinks" />
+        <Navigation
+            v-if="authService.isLoggedIn()"
+            :primary="primaryLinks"
+            :secondary="secondaryLinks"
+        />
         <v-main :class="backgroundClass">
             <RouterView />
             <ConfirmDialog ref="confirmDialog" />
@@ -9,7 +13,8 @@
 </template>
 
 <script setup lang="ts">
-import Navigation, { type NavigationLinks } from "@/components/MainNavigation.vue";
+import Navigation from "@/components/MainNavigation.vue";
+import type { NavigationLinks } from "@/components/NavigationList.vue";
 import { PermissionTarget, useAuth } from "@/services/auth";
 import { computed, provide, ref } from "vue";
 import { useI18n } from "vue-i18n";
@@ -20,24 +25,20 @@ const { t } = useI18n();
 const backgroundClass = computed(() => (authService.isLoggedIn() ? undefined : "login-view"));
 const authService = useAuth();
 const confirmDialog = ref(undefined);
-const navLinks: NavigationLinks = [
+const primaryLinks: NavigationLinks = [
     {
         title: t("App.nav.dashboard"),
         icon: "fas fa-dashboard",
         to: "dashboard",
     },
     {
-        title: t("App.nav.site"),
-        icon: "fas fa-sitemap",
-        children: [
-            {
-                title: t("App.nav.blog"),
-                icon: "fas fa-blog",
-                to: "blog",
-                requirePerm: PermissionTarget.BLOG_ENTRY,
-            },
-        ],
+        title: t("App.nav.blog"),
+        icon: "fas fa-blog",
+        to: "blog",
+        requirePerm: PermissionTarget.BLOG_ENTRY,
     },
+];
+const secondaryLinks: NavigationLinks = [
     {
         title: t("App.nav.logout"),
         icon: "fas fa-right-from-bracket",
